Trait memory::alloc::Allocator

pub trait Allocator {
    unsafe fn allocate(&mut self, size: usize, align: usize) -> Option<*mut u8>;
    unsafe fn deallocate(&mut self, frame: *mut u8, size: usize, align: usize);

    unsafe fn reallocate(
        &mut self, 
        old_frame: *mut u8, 
        old_size: usize, 
        new_size: usize, 
        align: usize
    ) -> Option<*mut u8> { ... }
    unsafe fn zero_alloc(
        &mut self, 
        _size: usize, 
        _align: usize
    ) -> Option<*mut u8> { ... }
}

An Allocator implements a particular memory allocation strategy.

Required Methods

unsafe fn allocate(&mut self, size: usize, align: usize) -> Option<*mut u8>

Allocate a new block of size size on alignment align.

Arguments:

size: the amount of memory to allocate (in bytes)

align: the alignment for the allocation request

Returns:

Some(*mut u8) if the request was allocated successfully

None if the allocator is out of memory or if the request was invalid.

Unsafe Because:

The allocated block must be deallocated manually.

unsafe fn deallocate(&mut self, frame: *mut u8, size: usize, align: usize)

Release an allocated block of memory.

The size and align parameters must be the same as the original size and alignment of the frame being deallocated, otherwise our heap will become corrupted.

Arguments:

frame: a pointer to the block of memory to deallocate

size: the size of the block being deallocated

align: the alignment of the block being deallocated

Unsafe Because:

Manually deallocates a block.

If the size and align parameters do not match the original layout of the requested block, the heap will become corrupted.

Provided Methods

unsafe fn reallocate(
&mut self,
old_frame: *mut u8,
old_size: usize,
new_size: usize,
align: usize
) -> Option<*mut u8>

Reallocate old_frame from old_size bytes to new_size bytes

The old_size and align parameters must be the same as the original size and alignment of the frame being reallocated, otherwise our heap will become corrupted.

Arguments:

old_frame: a pointer to the frame to be reallocated

old_size: the size (in bytes) of the frame being reallocated

new_size: the size to reallocate the frame to.

align: the alignment for the allocation request

Returns:

Some(*mut u8) if the frame was reallocated successfully

None if the allocator is out of memory or if the request was invalid.

unsafe fn zero_alloc(&mut self, _size: usize, _align: usize) -> Option<*mut u8>