Function memory::allocate_pages_by_bytes_at

pub fn allocate_pages_by_bytes_at(
    vaddr: VirtualAddress, 
    num_bytes: usize
) -> Result<AllocatedPages, &'static str>

Allocates pages starting at the given VirtualAddress with a size given in number of bytes.

This function still allocates whole pages by rounding up the number of bytes. See allocate_pages_deferred() for more details.