Function memory::init

pub fn init(
    boot_info: &BootInformation
) -> Result<(PageTable, NoDrop<MappedPages>, NoDrop<MappedPages>, NoDrop<MappedPages>, (AllocatedPages, NoDrop<MappedPages>), MappedPages, [Option<NoDrop<MappedPages>>; 32], [Option<NoDrop<MappedPages>>; 32]), &'static str>

Initializes the virtual memory management system. Consumes the given BootInformation, because after the memory system is initialized, the original BootInformation will be unmapped and inaccessible.

Returns the following tuple, if successful:

the kernel’s new PageTable, which is now currently active,

the MappedPages of the kernel’s text section,

the MappedPages of the kernel’s rodata section,

the MappedPages of the kernel’s data section,

a tuple of the stack’s underlying guard page (an AllocatedPages instance) and the actual MappedPages backing it,

the MappedPages holding the bootloader info,

the kernel’s list of other higher-half MappedPages that needs to be converted to a vector after heap initialization, and which should be kept forever,

the kernel’s list of identity-mapped MappedPages that needs to be converted to a vector after heap initialization, and which should be dropped before starting the first userspace program.