Trait memory::paging::Mapper

pub trait Mapper {
    type Flags;
    type Frame;
    fn translate(&self, vaddr: VAddr) -> Option<PAddr>;
    fn translate_page(&self, page: Page) -> Option<Self::Frame>;
    fn map_to<A>(
        &mut self, 
        page: Page, 
        frame: Self::Frame, 
        flags: Self::Flags, 
        alloc: &mut A
    )
    where
        A: FrameAllocator<Self::Frame>;
    fn identity_map<A>(
        &mut self, 
        frame: Self::Frame, 
        flags: Self::Flags, 
        alloc: &mut A
    )
    where
        A: FrameAllocator<Self::Frame>;
    fn map_to_any<A>(&mut self, page: Page, flags: Self::Flags, alloc: &mut A)
    where
        A: FrameAllocator<Self::Frame>;
}

Associated Types

type Flags

type Frame

Required Methods

fn translate(&self, vaddr: VAddr) -> Option<PAddr>

Translates a virtual address to the corresponding physical address.

Return

Some(PAddr) containing the physical address corresponding to vaddr, if it is mapped.

None: if the address is not mapped.

fn translate_page(&self, page: Page) -> Option<Self::Frame>

Translates a virtual page to a physical frame.

fn map_to<A>(
&mut self,
page: Page,
frame: Self::Frame,
flags: Self::Flags,
alloc: &mut A
) where
A: FrameAllocator<Self::Frame>,

Modifies the page tables so that page maps to frame.

Arguments

page: the virtual Page to map

frame: the physical Frame that Page should map to.

flags: the page table entry flags.

alloc: a memory allocator

fn identity_map<A>(
&mut self,
frame: Self::Frame,
flags: Self::Flags,
alloc: &mut A
) where
A: FrameAllocator<Self::Frame>,

Identity map a given frame.

Arguments

frame: the physical Frame to identity map

flags: the page table entry flags.

alloc: a memory allocator

fn map_to_any<A>(&mut self, page: Page, flags: Self::Flags, alloc: &mut A) where
A: FrameAllocator<Self::Frame>,

Map the given page to any free frame.

This is like the fire and forget version of map_to: we just pick the first available free frame and map the page to it.

Arguments

page: the virtual Page to map

flags: the page table entry flags.

alloc: a memory allocator