Struct memory::AllocatedFrames

pub struct AllocatedFrames { /* private fields */ }

Represents a range of allocated physical memory Frames; derefs to FrameRange.

These frames are not immediately accessible because they’re not yet mapped by any virtual memory pages. You must do that separately in order to create a MappedPages type, which can then be used to access the contents of these frames.

This object represents ownership of the range of allocated physical frames; if this object falls out of scope, its allocated frames will be auto-deallocated upon drop.

Implementations

impl AllocatedFrames

pub const fn empty() -> AllocatedFrames

Returns an empty AllocatedFrames object that performs no frame allocation. Can be used as a placeholder, but will not permit any real usage.

pub fn merge(&mut self, other: AllocatedFrames) -> Result<(), AllocatedFrames>

Merges the given AllocatedFrames object other into this AllocatedFrames object (self). This is just for convenience and usability purposes, it performs no allocation or remapping.

The given other must be physically contiguous with self, i.e., come immediately before or after self. That is, either self.start == other.end + 1 or self.end + 1 == other.start must be true.

If either of those conditions are met, self is modified and Ok(()) is returned, otherwise Err(other) is returned.

pub fn split(
self,
at_frame: Frame
) -> Result<(AllocatedFrames, AllocatedFrames), AllocatedFrames>

Splits this AllocatedFrames into two separate AllocatedFrames objects:

[beginning : at_frame - 1]

[at_frame : end]

This function follows the behavior of core::slice::split_at(), thus, either one of the returned AllocatedFrames objects may be empty.

If at_frame == self.start, the first returned AllocatedFrames object will be empty.

If at_frame == self.end + 1, the second returned AllocatedFrames object will be empty.

Returns an Err containing this AllocatedFrames if at_frame is otherwise out of bounds.

pub fn as_allocated_frame(&self) -> AllocatedFrame<'_>

Returns an AllocatedFrame if this AllocatedFrames object contains only one frame.

Panic

Panics if this AllocatedFrame contains multiple frames or zero frames.