Struct memory::BorrowedMappedPages

pub struct BorrowedMappedPages<T, M = Immutable, B = MappedPages>where
    T: FromBytes,
    M: Mutability,
    B: Borrow<MappedPages>,{ /* private fields */ }

A borrowed MappedPages object that derefs to &T and optionally also &mut T.

Type parameters

T: FromBytes: the same parameter used in MappedPages::as_type() functions.

M: Mutability: whether or not the MappedPages can be mutably borrowed.

For ease of use, the default is Immutable.

B: Borrow<MappedPages>: enables the user of this type to use a borrow-able wrapper type or reference around MappedPages.

For ease of use, the default is just a plain MappedPages with no wrapper. This default MappedPages type can be borrowed immutably or mutably.

Example types that satisfy the bounds for B are Arc<MappedPages, Rc<MappedPages>, &MappedPages, etc, all of which can only be borrowed immutably via the from() method, as these types do not permit mutable access. For mutable borrows, see the from_mut() method, which works with types like MappedPages itself or &mut MappedPages.

Drop behavior

When dropped, the borrow ends and the contained MappedPages is dropped and unmapped.

Also, you can manually end the borrow to reclaim the inner B MappedPages type via the Self::into_inner() method.

Implementations

impl<T: FromBytes, B: Borrow<MappedPages>> BorrowedMappedPages<T, Immutable, B>

pub fn from(mp: B, byte_offset: usize) -> Result<Self, (B, &'static str)>

Immutably borrows the given MappedPages as an instance of type &T located at the given byte_offset into the MappedPages.

See MappedPages::as_type() for more info.

Arguments

mp: the MappedPages that you wish to immutably borrow as an instance of &T.

See the type-level docs for more info and examples of how to use this argument.

byte_offset: the offset (in number of bytes) from the beginning of the MappedPages memory region at which the struct T is located (where it should start).

This offset must be properly aligned with respect to the alignment requirements of type T, otherwise an error will be returned.

Upon failure, this returns an error tuple containing the unmodified mp argument and a string describing the error.

impl<T: FromBytes, B: BorrowMut<MappedPages>> BorrowedMappedPages<T, Mutable, B>

pub fn from_mut(mp: B, byte_offset: usize) -> Result<Self, (B, &'static str)>

Mutably borrows the given MappedPages as an instance of type &mut T located at the given byte_offset into the MappedPages.

See MappedPages::as_type_mut() for more info.

Arguments

mp: the MappedPages that you wish to mutably borrow as an instance of &mut T.

See the type-level docs for more info and examples of how to use this argument.

byte_offset: the offset (in number of bytes) from the beginning of the MappedPages memory region at which the struct T is located (where it should start).

This offset must be properly aligned with respect to the alignment requirements of type T, otherwise an error will be returned.

Upon failure, returns an error containing the unmodified MappedPages and a string describing the error.

impl<T: FromBytes, M: Mutability, B: Borrow<MappedPages>> BorrowedMappedPages<T, M, B>

pub fn into_inner(self) -> B

Consumes this object and returns the inner MappedPages value (more specifically, the Borrow-able container holding the MappedPages).

pub fn inner_ref(&self) -> &B

Returns a reference to the inner MappedPages value (more specifically, the Borrow-able container holding the MappedPages).