Struct memory::MappedPages

pub struct MappedPages { /* fields omitted */ }

Represents a contiguous range of virtual memory pages that are currently mapped. A MappedPages object can only have a single range of contiguous pages, not multiple disjoint ranges. This does not guarantee that its pages are mapped to frames that are contiguous in physical memory.

This object also represents ownership of those pages; if this object falls out of scope, it will be dropped, and the pages will be unmapped, and if they were allocated, then also de-allocated. Thus, it ensures memory safety by guaranteeing that this object must be held in order to access data stored in these mapped pages, just like a MutexGuard guarantees that data protected by a Mutex can only be accessed while that Mutex's lock is held.

Methods

impl MappedPages

pub fn empty() -> MappedPages

Returns an empty MappedPages object that performs no allocation or mapping actions. Can be used as a placeholder, but will not permit any real usage.

pub fn flags(&self) -> EntryFlags

Returns the flags that describe this MappedPages page table permissions.

pub fn from_existing(
already_mapped_pages: PageRange,
flags: EntryFlags
) -> MappedPages

Deprecated

Constructs a MappedPages object from an already existing mapping. Useful for creating idle task Stacks, for example.

pub fn merge(
&mut self,
mp: MappedPages
) -> Result<(), (&'static str, MappedPages)>

Merges the given MappedPages object into this MappedPages object.

For example, if you have the following MappedPages objects:

this mapping, with a page range including one page at 0x2000

mp, with a page range including two pages at 0x3000 and 0x4000 Then this MappedPages object will be updated to cover three pages from [0x2000:0x4000] inclusive.

In addition, the MappedPages objects must have the same flags and page table root frame (i.e., they must have all been mapped using the same set of page tables).

In addition, the MappedPages objects must either all have AllocatedPages or all have no AllocatedPages. MappedPages that were mapped to allocated virtual pages cannot be merged with those that weren't mapped to allocated pages.

If an error occurs, such as the mappings not being contiguous or having different flags, then a tuple including an error message and the original mp will be returned, which prevents the mp from being dropped.

Note

No remapping actions or page reallocations will occur on either a failure or a success.

pub fn deep_copy<A: FrameAllocator>(
&self,
new_flags: Option<EntryFlags>,
active_table_mapper: &mut Mapper,
allocator: &mut A
) -> Result<MappedPages, &'static str>

Creates a deep copy of this MappedPages memory region, by duplicating not only the virtual memory mapping but also the underlying physical memory frames.

The caller can optionally specify new flags for the duplicated mapping, otherwise, the same flags as the existing MappedPages will be used. This is useful for when you want to modify contents in the new pages, since it avoids extra remap() operations.

Returns a new MappedPages object with the same in-memory contents as this object, but at a completely new memory region.

pub fn remap(
&mut self,
active_table_mapper: &mut Mapper,
new_flags: EntryFlags
) -> Result<(), &'static str>

Change the permissions (new_flags) of this MappedPages's page table entries.

pub fn as_type<T>(&self, offset: usize) -> Result<&T, &'static str>

Reinterprets this MappedPages's underlying memory region as a struct of the given type, i.e., overlays a struct on top of this mapped memory region.

Arguments

offset: the offset into the memory region at which the struct is located (where it should start).

Returns a reference to the new struct (&T) that is formed from the underlying memory region, with a lifetime dependent upon the lifetime of this MappedPages object. This ensures safety by guaranteeing that the returned struct reference cannot be used after this MappedPages object is dropped and unmapped.

pub fn as_type_mut<T>(&mut self, offset: usize) -> Result<&mut T, &'static str>

Same as as_type(), but returns a mutable reference to the type T.

Thus, it checks to make sure that the underlying mapping is writable.

pub fn as_slice<T>(
&self,
byte_offset: usize,
length: usize
) -> Result<&[T], &'static str>

Reinterprets this MappedPages's underlying memory region as a slice of any type.

Arguments

byte_offset: the offset (in number of bytes) into the memory region at which the slice should start.

length: the length of the slice, i.e., the number of T elements in the slice. Thus, the slice will go from offset to offset + (sizeof(T) * length).

Returns a reference to the new slice that is formed from the underlying memory region, with a lifetime dependent upon the lifetime of this MappedPages object. This ensures safety by guaranteeing that the returned slice cannot be used after this MappedPages object is dropped and unmapped.

pub fn as_slice_mut<T>(
&mut self,
byte_offset: usize,
length: usize
) -> Result<&mut [T], &'static str>

Same as as_slice(), but returns a mutable slice.

Thus, it checks to make sure that the underlying mapping is writable.

pub fn as_func<'a, F>(
&self,
offset: usize,
space: &'a mut usize
) -> Result<&'a F, &'static str>

Reinterprets this MappedPages's underlying memory region as an executable function with any signature.

Arguments

offset: the offset (in number of bytes) into the memory region at which the function starts.

space: a hack to satisfy the borrow checker's lifetime requirements.

Returns a reference to the function that is formed from the underlying memory region, with a lifetime dependent upon the lifetime of the given space object.

TODO FIXME: ideally, we'd have an integrated function that checks with the mod_mgmt crate to see if the size of the function can fit (not just the size of the function POINTER, which will basically always fit) within the bounds of this MappedPages object; this integrated function would be based on the given string name of the function, like "task::this::foo", and would invoke this as_func() function directly.

We have to accept space for the function pointer to exist, because it cannot live in this function's stack. It has to live in stack of the function that invokes the actual returned function reference, otherwise there would be a lifetime issue and a guaranteed page fault. So, the space arg is a hack to ensure lifetimes; we don't care about the actual value of space, as the value will be overwritten, and it doesn't matter both before and after the call to this as_func().

The generic F parameter is the function type signature itself, e.g., fn(String) -> u8.

Examples

Here's how you might call this function:

type PrintFuncSignature = fn(&str) -> Result<(), &'static str>;
let mut space = 0; // this must persist throughout the print_func being called
let print_func: &PrintFuncSignature = mapped_pages.as_func(func_offset, &mut space).unwrap();
print_func("hi");

Because Rust has lexical lifetimes, the space variable must have a lifetime at least as long as the print_func variable, meaning that space must still be in scope in order for print_func to be invoked.