Struct memory::Mapper

pub struct Mapper {
    pub target_p4: Frame,
    // some fields omitted
}

Fields
target_p4: Frame

The Frame contaning the top-level P4 page table.

Methods

impl Mapper

pub fn from_current() -> Mapper

pub fn with_p4_frame(p4: Frame) -> Mapper

pub fn p4(&self) -> &Table<Level4>

pub fn p4_mut(&mut self) -> &mut Table<Level4>

pub fn dump_pte(&self, virtual_address: VirtualAddress)

Dumps all page table entries at all four levels for the given VirtualAddress, and also shows their EntryFlags.

Useful for debugging page faults.

pub fn translate(
&self,
virtual_address: VirtualAddress
) -> Option<PhysicalAddress>

Translates a VirtualAddress to a PhysicalAddress by walking the page tables.

pub fn translate_page(&self, page: Page) -> Option<Frame>

Translates a virtual memory Page to a physical memory Frame by walking the page tables.

pub fn map_to<A>(
&mut self,
page: Page,
frame: Frame,
flags: EntryFlags,
allocator: &mut A
) -> Result<MappedPages, &'static str> where
A: FrameAllocator,

creates a mapping for a specific page -> specific frame

pub fn map<A>(
&mut self,
page: Page,
flags: EntryFlags,
allocator: &mut A
) -> Result<MappedPages, &'static str> where
A: FrameAllocator,

maps the given Page to a randomly selected (newly allocated) Frame

pub fn map_pages<A>(
&mut self,
pages: PageRange,
flags: EntryFlags,
allocator: &mut A
) -> Result<MappedPages, &'static str> where
A: FrameAllocator,

maps the given Pages to a randomly selected (newly allocated) Frame

pub fn map_frames<A>(
&mut self,
frames: FrameRange,
start_page: Page,
flags: EntryFlags,
allocator: &mut A
) -> Result<MappedPages, &'static str> where
A: FrameAllocator,

maps the given contiguous range of Frames frame_range to contiguous Pages starting at start_page

pub fn map_allocated_pages_to<A>(
&mut self,
allocated_pages: AllocatedPages,
frames: FrameRange,
flags: EntryFlags,
allocator: &mut A
) -> Result<MappedPages, &'static str> where
A: FrameAllocator,

maps the given AllocatedPages to the given actual frames. Consumes the given AllocatedPages and returns a MappedPages object which contains that AllocatedPages object.

pub fn map_allocated_pages<A>(
&mut self,
allocated_pages: AllocatedPages,
flags: EntryFlags,
allocator: &mut A
) -> Result<MappedPages, &'static str> where
A: FrameAllocator,

maps the given AllocatedPages to randomly chosen (allocated) frames Consumes the given AllocatedPages and returns a MappedPages object which contains that AllocatedPages object.