Struct memory::Page

pub struct Page { /* private fields */ }

A Page is a chunk of virtual memory aligned to a PAGE_SIZE boundary.

Implementations

impl Page

pub const fn start_address(&self) -> VirtualAddress

Returns the VirtualAddress at the start of this Page.

pub const fn number(&self) -> usize

Returns the number of this Page.

pub const fn containing_address(addr: VirtualAddress) -> Page

Returns the Page containing the given VirtualAddress.

impl Page

pub const fn p4_index(&self) -> usize

Returns the 9-bit part of this Page’s VirtualAddress that is the index into the P4 page table entries list.

pub const fn p3_index(&self) -> usize

Returns the 9-bit part of this Page’s VirtualAddress that is the index into the P3 page table entries list.

pub const fn p2_index(&self) -> usize

Returns the 9-bit part of this Page’s VirtualAddress that is the index into the P2 page table entries list.

pub const fn p1_index(&self) -> usize

Returns the 9-bit part of this Page’s VirtualAddress that is the index into the P1 page table entries list.

Using this returned usize value as an index into the P1 entries list will give you the final PTE, from which you can extract the mapped Frame using PageTableEntry::pointed_frame().