Struct memory::PageRange

pub struct PageRange(_);

A range of Pages that are contiguous in virtual memory.

Methods

impl PageRange

pub const fn new(start: Page, end: Page) -> PageRange

Creates a new range of Pages that spans from start to end, both inclusive bounds.

pub const fn empty() -> PageRange

Creates a PageRange that will always yield None.

pub fn from_virt_addr(
starting_virt_addr: VirtualAddress,
size_in_bytes: usize
) -> PageRange

A convenience method for creating a new PageRange that spans all Pages from the given virtual address to an end bound based on the given size.

pub const fn start_address(&self) -> VirtualAddress

Returns the VirtualAddress of the starting Page.

pub const fn size_in_pages(&self) -> usize

Returns the size in number of Pages. Use this instead of the Iterator trait's count() method. This is instant, because it doesn't need to iterate over each Page, unlike normal iterators.

pub const fn size_in_bytes(&self) -> usize

Returns the size in number of bytes.

pub fn contains_virt_addr(&self, virt_addr: VirtualAddress) -> bool

Whether this PageRange contains the given VirtualAddress.

pub fn offset_of_address(&self, virt_addr: VirtualAddress) -> Option<usize>

Returns the offset of the given VirtualAddress within this PageRange, i.e., the difference between virt_addr and self.start_address(). If the given VirtualAddress is not covered by this range of Pages, this returns None.

Examples

If the page range covered addresses 0x2000 to 0x4000, then calling offset_of_address(0x3500) would return Some(0x1500).

pub fn address_at_offset(&self, offset: usize) -> Option<VirtualAddress>

Returns the VirtualAddress at the given offset into this mapping,
If the given offset is not covered by this range of Pages, this returns None.

Examples

If the page range covered addresses 0xFFFFFFFF80002000 to 0xFFFFFFFF80004000, then calling address_at_offset(0x1500) would return Some(0xFFFFFFFF80003500).