Struct memory::PageTable

pub struct PageTable { /* fields omitted */ }

A root (P4) page table.

Auto-derefs into a Mapper for easy invocation of memory mapping functions.

Methods

impl PageTable

pub fn new_table(
current_page_table: &mut PageTable,
new_p4_frame: Frame,
temporary_page: TemporaryPage
) -> Result<PageTable, &'static str>

Initializes a brand new top-level P4 PageTable (previously called an InactivePageTable) that is based on the given current_active_table and is located in the given new_p4_frame. The TemporaryPage is used for recursive mapping, and is auto-unmapped upon return.

Returns the new PageTable that exists in physical memory at the given new_p4_frame, and has the kernel memory region mappings copied in from the given current_page_table to ensure that the system will continue running

pub fn with<F>(
&mut self,
other_table: &mut PageTable,
temporary_page: TemporaryPage,
f: F
) -> Result<(), &'static str> where
F: FnOnce(&mut Mapper) -> Result<(), &'static str>,

Temporarily maps the given other PageTable to the recursive entry (510th entry) so that the given closure f can set up new mappings on the new other_table without actually switching to it yet. Accepts a closure f that is passed a Mapper, such that it can set up new mappings on the other table. Consumes the given temporary_page and automatically unmaps it afterwards.

Note

This does not perform any task switching or changing of the current page table register (e.g., cr3).

pub fn switch(&mut self, new_table: &PageTable) -> PageTable

Switches from the currently-active page table (this PageTable, i.e., self) to the given new_table. Returns the newly-switched-to PageTable.

pub fn physical_address(&self) -> PhysicalAddress

Returns the physical address of this page table's top-level p4 frame