Struct memory::PhysicalAddress

#[repr(transparent)]
pub struct PhysicalAddress(_);

A physical memory address, which is a usize under the hood.

Implementations

impl PhysicalAddress

pub fn new(addr: usize) -> Option<PhysicalAddress>

Creates a new PhysicalAddress, returning an error if the address is not canonical.

This is useful for checking whether an address is valid before using it. For example, on x86_64, virtual addresses are canonical if their upper bits (64:48] are sign-extended from bit 47, and physical addresses are canonical if their upper bits (64:52] are 0.

pub const fn new_canonical(addr: usize) -> PhysicalAddress

Creates a new PhysicalAddress that is guaranteed to be canonical.

pub const fn zero() -> PhysicalAddress

Creates a new PhysicalAddress with a value 0.

pub const fn value(&self) -> usize

Returns the underlying usize value for this PhysicalAddress.

pub const fn frame_offset(&self) -> usize

Returns the offset from the frame boundary specified by this `PhysicalAddress.

For example, if the PAGE_SIZE is 4096 (4KiB), then this will return the least significant 12 bits (12:0] of this PhysicalAddress.