Struct memory::PhysicalPage

pub struct PhysicalPage {
    pub number: u64,
}

A frame (physical page)

Fields
number: u64
Methods

impl PhysicalPage

const fn base_addr(&self) -> PAddr

Returns the physical address where this frame starts.

const fn containing_addr(addr: PAddr) -> PhysicalPage

Returns a new frame containing addr

unsafe fn as_ptr<T>(&self) -> *const T

Convert the frame into a raw pointer to the frame's base address

unsafe fn as_mut_ptr<T>(&self) -> *mut T

Convert the frame into a raw mutable pointer to the frame's base address