Struct memory::VirtualAddress

#[repr(transparent)]
pub struct VirtualAddress(_);

A virtual memory address, which is a usize under the hood.

Methods

impl VirtualAddress

pub fn new(virt_addr: usize) -> Result<VirtualAddress, &'static str>

Creates a new VirtualAddress, checking that the address is canonical, i.e., bits (64:48] are sign-extended from bit 47.

pub const fn new_canonical(virt_addr: usize) -> VirtualAddress

Creates a new VirtualAddress that is guaranteed to be canonical by forcing the upper bits (64:48] to be sign-extended from bit 47.

pub const fn zero() -> VirtualAddress

Creates a VirtualAddress with the value 0.

pub const fn value(&self) -> usize

Returns the underlying usize value for this VirtualAddress.

pub const fn page_offset(&self) -> usize

Returns the offset that this VirtualAddress specifies into its containing memory Page.

For example, if the PAGE_SIZE is 4KiB, then this will return the least significant 12 bits (12:0] of this VirtualAddress.