Trait memory::MemRange

pub trait MemRange {
    fn length(&self) -> usize;
    fn drop_front(&mut self, n: usize) -> &mut Self;
    fn drop_back(&mut self, n: usize) -> &mut Self;
    fn add_front(&mut self, n: usize) -> &mut Self;
    fn add_back(&mut self, n: usize) -> &mut Self;
}

Required Methods

fn length(&self) -> usize

Returns the number of Pages in this ranage

fn drop_front(&mut self, n: usize) -> &mut Self

Remove n pages from the beginning of this PageRange

fn drop_back(&mut self, n: usize) -> &mut Self

Remove n pages from the end of this PageRange

fn add_front(&mut self, n: usize) -> &mut Self

Add n pages at the front of this PageRange

fn add_back(&mut self, n: usize) -> &mut Self

Add n pages at the back of this PageRange