Trait memory::Page

pub trait Page where
    Self: Sized,
    Self: AddAssign<usize> + SubAssign<usize>,
    Self: Add<usize, Output = Self> + Sub<usize, Output = Self>,
    Self: PartialEq<Self> + PartialOrd<Self>,
    Self: Copy + Clone, {
    type Address: Addr;
    fn containing(addr: Self::Address) -> Self;
    fn base(&self) -> Self::Address;
    fn end_address(&self) -> Self::Address;
    fn number(&self) -> usize;

    fn range_of(&self, n: usize) -> Range<Self> { ... }
    fn range_until(&self, end: Self) -> Range<Self> { ... }
}

Trait for a page. These can be virtual pages or physical frames.

Associated Types

type Address: Addr

The type of address used to address this Page.

Typically, this is a PAddr or VAddr (but it could be a "MAddr") in schemes where we differentiate between physical and machine addresses. If we ever have those.

Required Methods

fn containing(addr: Self::Address) -> Self

Returns a new Page containing the given Address.

N.B. that since trait functions cannot be const, implementors of this trait may wish to provide implementations of this function outside of the impl block and then wrap them here.

fn base(&self) -> Self::Address

Returns the base Address where this page starts.

fn end_address(&self) -> Self::Address

Returns the end Address of this Page.

fn number(&self) -> usize

Provided Methods

fn range_of(&self, n: usize) -> Range<Self>

Returns a PageRange of this Page and the next n pages.

fn range_until(&self, end: Self) -> Range<Self>

Returns a PageRange on the frames from this frame until the end frame