Trait memory::PageSize

pub trait PageSize: Ord + PartialOrd<Self> + Clone + Copy + Sealed + 'static {
    const SIZE: MemChunkSize;
    const NUM_4K_PAGES: usize;
    const SIZE_IN_BYTES: usize;
}

Trait that represents the size of a page or frame, i.e., for normal or huge pages.

This is used to parameterize Page- and Frame-related types with a page size, in order to define normal and huge pages in a generic manner.

Required Associated Constants

const SIZE: MemChunkSize

const NUM_4K_PAGES: usize

const SIZE_IN_BYTES: usize