Function memory::init_post_heap

pub fn init_post_heap(
    page_table: PageTable,
    higher_half_mapped_pages: [Option<NoDrop<MappedPages>>; 32],
    identity_mapped_pages: [Option<NoDrop<MappedPages>>; 32],
    heap_mapped_pages: MappedPages
) -> (MmiRef, NoDrop<Vec<MappedPages>>)

Finishes initializing the memory management system after the heap is ready.

Returns the following tuple:

The kernel’s new MemoryManagementInfo, representing the initial virtual address space,

The kernel’s list of identity-mapped MappedPages, which must not be dropped until all AP (additional CPUs) are fully booted, but should be dropped before starting the first user application.