Function memory::init_post_heap

pub fn init_post_heap(
    page_table: PageTable,
    higher_half_mapped_pages: [Option<MappedPages>; 32],
    identity_mapped_pages: [Option<MappedPages>; 32],
    heap_mapped_pages: MappedPages
) -> Result<(MmiRef, Vec<MappedPages>), &'static str>

Finishes initializing the virtual memory management system after the heap is initialized and returns a MemoryManagementInfo instance, which represents the initial (the kernel’s) address space.

Returns the following tuple, if successful:

The kernel’s new MemoryManagementInfo

The kernel’s list of identity-mapped MappedPages which should be dropped before starting the first userspace program.