Struct memory::BorrowedMappedPages

pub struct BorrowedMappedPages<T: FromBytes, M: Mutability = Immutable> { /* private fields */ }

A borrowed MappedPages object that derefs to &T and optionally also &mut T.

By default, the Mutability type parameter is Immutable for ease of use.

When dropped, the borrow ends and the contained MappedPages is dropped and unmapped. You can manually end the borrow and reclaim the inner MappedPages via Self::into_inner().

Implementations

impl<T: FromBytes> BorrowedMappedPages<T, Immutable>

pub fn from(
mp: MappedPages,
byte_offset: usize
) -> Result<BorrowedMappedPages<T>, (MappedPages, &'static str)>

Immutably borrows the given MappedPages as an instance of type &T located at the given byte_offset into the MappedPages.

See MappedPages::as_type() for more info.

Upon failure, returns an error containing the unmodified MappedPages and a string describing the error.

impl<T: FromBytes> BorrowedMappedPages<T, Mutable>

pub fn from_mut(
mp: MappedPages,
byte_offset: usize
) -> Result<BorrowedMappedPages<T, Mutable>, (MappedPages, &'static str)>

Mutably borrows the given MappedPages as an instance of type &mut T located at the given byte_offset into the MappedPages.

See MappedPages::as_type_mut() for more info.

Upon failure, returns an error containing the unmodified MappedPages and a string describing the error.

impl<T: FromBytes, M: Mutability> BorrowedMappedPages<T, M>

pub fn into_inner(self) -> MappedPages

Consumes this object and returns the inner MappedPages.