Struct memory::Mapper

pub struct Mapper { /* fields omitted */ }

Implementations

impl Mapper

pub fn dump_pte<W: Write>(
&self,
writer: &mut W,
virtual_address: VirtualAddress
) -> Result

Dumps all page table entries at all four page table levels for the given VirtualAddress, and also shows their EntryFlags.

The page table details are written to the the given writer.

pub fn translate(
&self,
virtual_address: VirtualAddress
) -> Option<PhysicalAddress>

Translates a VirtualAddress to a PhysicalAddress by walking the page tables.

pub fn translate_page(&self, page: Page) -> Option<Frame>

Translates a virtual memory Page to a physical memory Frame by walking the page tables.

pub fn map_allocated_pages_to(
&mut self,
pages: AllocatedPages,
frames: AllocatedFrames,
flags: EntryFlags
) -> Result<MappedPages, &'static str>

Maps the given virtual AllocatedPages to the given physical AllocatedFrames.

Consumes the given AllocatedPages and returns a MappedPages object which contains those AllocatedPages.

pub fn map_allocated_pages(
&mut self,
pages: AllocatedPages,
flags: EntryFlags
) -> Result<MappedPages, &'static str>

Maps the given AllocatedPages to randomly chosen (allocated) physical frames.

Consumes the given AllocatedPages and returns a MappedPages object which contains those AllocatedPages.