Struct memory::PageTable

pub struct PageTable { /* fields omitted */ }

A top-level root (P4) page table.

Auto-derefs into a Mapper for easy invocation of memory mapping functions.

Implementations

impl PageTable

pub fn new_table(
current_page_table: &mut PageTable,
new_p4_frame: AllocatedFrames,
temporary_page: TemporaryPage
) -> Result<PageTable, &'static str>

Initializes a brand new top-level P4 PageTable whose root is located in the given new_p4_frame. It requires using the given current_active_table to set up its initial mapping contents, for which the given TemporaryPage is used to recursively map it (and is auto-unmapped upon return).

Returns the new PageTable that exists in physical memory at the given new_p4_frame. Note that this new page table has no current mappings beyond the recursive P4 mapping, so you will need to create or copy over any relevant mappings before using (switching) to this new page table in order to ensure the system keeps running.

pub fn with<F>(
&mut self,
other_table: &mut PageTable,
temporary_page: TemporaryPage,
f: F
) -> Result<(), &'static str> where
F: FnOnce(&mut Mapper) -> Result<(), &'static str>,

Temporarily maps the given other PageTable to the recursive entry (510th entry) so that the given closure f can set up new mappings on the new other_table without actually switching to it yet. Accepts a closure f that is passed a Mapper, such that it can set up new mappings on the other table. Consumes the given temporary_page and automatically unmaps it afterwards.

Note

This does not perform any task switching or changing of the current page table register (e.g., cr3).

pub fn switch(&mut self, new_table: &PageTable)

Switches from the currently-active page table (this PageTable, i.e., self) to the given new_table. After this function, the given new_table will be the currently-active PageTable.

pub fn physical_address(&self) -> PhysicalAddress

Returns the physical address of this page table’s top-level p4 frame